Backend/requirements.txt
langchain
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
fastapi
chromadb
langchain-chroma
unstructured
openai
ddgs
unstructured
supabase
apscheduler
uvicorn
dateparser
playwright
pytesseract
pyowm
